+++ app/common/infrastructure/mq/listeners/stomp_listener_base.py
@@ -8,7 +8,7 @@ from abc import abstractmethod, ABC
 import stomp
 from stomp.utils import Frame
 
-from app.ingest.infrastructure.mq.stomp_interactor import StompInteractor
+from app.common.infrastructure.mq.stomp_interactor import StompInteractor
 
 
 class StompListenerBase(stomp.ConnectionListener, StompInteractor, ABC):
