+++ app/ingest/infrastructure/mq/listeners/process_status_queue_listener.py
@@ -10,8 +10,8 @@ from app.containers import Services
 from app.ingest.domain.services.exceptions.get_ingest_by_package_id_exception import GetIngestByPackageIdException
 from app.ingest.domain.services.exceptions.set_ingest_as_processed_exception import SetIngestAsProcessedException
 from app.ingest.domain.services.ingest_service import IngestService
-from app.ingest.infrastructure.mq.listeners.stomp_listener_base import StompListenerBase
-from app.ingest.infrastructure.mq.mq_connection_params import MqConnectionParams
+from app.common.infrastructure.mq.listeners.stomp_listener_base import StompListenerBase
+from app.common.infrastructure.mq.mq_connection_params import MqConnectionParams
 
 
 class ProcessStatusQueueListener(StompListenerBase):
