+++ app/ingest/infrastructure/mq/publishers/process_ready_queue_publisher.py
@@ -7,8 +7,8 @@ import os
 
 from app.ingest.domain.models.ingest.ingest import Ingest
 from app.ingest.domain.mq.process_ready_queue_publisher import IProcessReadyQueuePublisher
-from app.ingest.infrastructure.mq.mq_connection_params import MqConnectionParams
-from app.ingest.infrastructure.mq.publishers.stomp_publisher_base import StompPublisherBase
+from app.common.infrastructure.mq.mq_connection_params import MqConnectionParams
+from app.common.infrastructure.mq.publishers.stomp_publisher_base import StompPublisherBase
 
 
 class ProcessReadyQueuePublisher(IProcessReadyQueuePublisher, StompPublisherBase):
